fix: autorise les valeurs décimales pour les frais de ménage et d'indemnisation
- Modifie le type des propriétés de coût de Int à Double pour plus de précision.
- Met à jour la logique de parsing des montants pour supporter les nombres à virgule.

+++ RentUp/Controllers/Bookings/Edit/RU_Bookings_Edit_ViewController.swift
@@ -279,7 +279,7 @@ public class RU_Bookings_Edit_ViewController : RU_ViewController {
             
             if let value = self?.costsCleaningTextFieldRowStack.textField.text {
                 
-                self?.booking?.costs.cleaning = Int(value)
+                self?.booking?.costs.cleaning = Double(value)
                 self?.updateSaveButton()
             }
             
@@ -299,7 +299,7 @@ public class RU_Bookings_Edit_ViewController : RU_ViewController {
             
             if let value = self?.costsCompensationTextFieldRowStack.textField.text {
                 
-                self?.booking?.costs.compensation = Int(value)
+                self?.booking?.costs.compensation = Double(value)
                 self?.updateSaveButton()
             }
